+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Automatically create compressed copies of files with `br` and `gzip` using new `autoCompress` option.
+- Automatically create compressed copies of files with `br` and `gzip` using new `encoding` option.
 - Object tagging for AWS and Azure.
 
 ## [3.0.0] - 2020-04-27